feature: Add loading status when deleting comments.

+++ src/app/posts/[id]/_components/comments/CommentDelete.tsx
@@ -1,5 +1,7 @@
 import DeleteIcon from "@/components/icons/DeleteIcon";
+import SixDotsRotate from "@/components/icons/SixDotsRotate";
 import { useQueryClient } from "@tanstack/react-query";
+import { useState } from "react";
 
 export default function CommentDelete({
   postId,
@@ -9,22 +11,40 @@ export default function CommentDelete({
   commentId: string;
 }) {
   const queryClient = useQueryClient();
+  const [loading, setLoading] = useState<boolean>(false);
   const handleOnClick = async () => {
-    
-    const response = await fetch(`/api/posts/${postId}/comments/${commentId}`, {
-      method: "DELETE",
-    });
-
-    if (response.ok) {
-      queryClient.invalidateQueries({
-        queryKey: [`${postId}/comments`],
-      });
-    } else return;
+    const confirm = window.confirm("Delete this comment?");
+    if (!confirm) return;
+    try {
+      setLoading(true);
+      const response = await fetch(
+        `/api/posts/${postId}/comments/${commentId}`,
+        {
+          method: "DELETE",
+        },
+      );
+      if (response.ok) {
+        queryClient.invalidateQueries({
+          queryKey: [`${postId}/comments`],
+        });
+      }
+    } catch (e) {
+      console.log(e);
+      alert("Failed to delete the comment. Try again.");
+    } finally {
+      setLoading(false);
+    }
   };
 
   return (
-    <button id={commentId} onClick={handleOnClick}>
-      <DeleteIcon className="w-6 hover:stroke-red-500 cursor-pointer" />
-    </button>
+    <>
+      {loading ? (
+        <SixDotsRotate className="w-6" />
+      ) : (
+        <button id={commentId} onClick={handleOnClick}>
+          <DeleteIcon className="w-6 hover:stroke-red-500 cursor-pointer" />
+        </button>
+      )}
+    </>
   );
 }
